Remove legacy `BomUploadProcessingTask`
* Makes the new implementation (previously known as `BomUploadProcessingTaskV2`) the default
* Removes code related to the legacy implementation
* Removes the `experimental:bom.processing.task.v2.enabled` property
