+++ src/main/java/org/dependencytrack/persistence/QueryManager.java
@@ -914,10 +914,6 @@ public class QueryManager extends AlpineQueryManager {
         return getVulnerableSoftwareQueryManager().getAllVulnerableSoftware(cpePart, cpeVendor, cpeProduct, purl);
     }
 
-    public Component matchSingleIdentity(final Project project, final ComponentIdentity cid) {
-        return getComponentQueryManager().matchSingleIdentity(project, cid);
-    }
-
     public Component matchSingleIdentityExact(final Project project, final ComponentIdentity cid) {
         return getComponentQueryManager().matchSingleIdentityExact(project, cid);
     }
